app/requirements.txt
psycopg2-binary
streamlit

sentence_transformers==2.7.0
openai
numpy==1.24.1
elasticsearch==8.4.3 
